# Backend-Javascript/Platinum/chapter 1/MVC/article/docs/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Swagger Article and Item - OpenAPI 3.0",
    "contact": {
      "email": "erandadava@gmail.com"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "version": "1.0.11"
  },
  "servers": [
    {
      "url": "http://localhost:3000/api/v0"
    }
  ],
...
  "paths": {
...
    "/item/{itemId}":{
      "put": {
        "tags": [
          "Item"
        ],
        "summary": "Update an Item",
        "description": "Update an Item",
        "operationId": "UpdateItem",
        "parameters":[{
          "name": "itemId",
          "in": "path",
          "description": "id of Item",
          "required": true,
          "schema":{
            "type": "integer"
          }
        }],
        "requestBody": {
          "description": "Update an Item",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/addItem"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccesfullAddItem"
                }
              }
            }
          },
          "405": {
            "description": "Invalid input"
          }
        }
      },
      "delete": {
        "tags": [
          "Item"
        ],
        "summary": "Delete an Item",
        "description": "delete an Item",
        "operationId": "deleteItem",
        "parameters": [
          {
            "name": "itemId",
            "in": "path",
            "description": "id of Item",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid Item value"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "addItem": {
        "type": "object",
        "properties":{
          "namaItem":{
            "type": "string",
            "format": "int64",
            "example": "Produk produk indonesia"
          },
          "qty": {
            "type": "integer",
            "example": 2
          },
          "approved": {
            "type": "boolean",
            "example": true
          }
        }
      },
      "SuccesfullAddItem": {
        "type":"object",
        "properties":{
          "status":{
            "type": "integer",
            "format": "int64",
            "example": 200
          },
          "message":{
            "type": "string",
            "example": "Success"
          },
          "response":{
            "type":"object",
            "properties": {
              "id" :{
                "type":"integer",
                "example":""
              },
              "namaItem" :{
                "type":"string",
                "example":""
              },
              "qty" :{
                "type":"integer",
                "example":""
              },
              "approved" :{
                "type":"boolean",
                "example":""
              },
              "createdAt" :{
                "type":"string",
                "example":""
              },
              "updatedAt" :{
                "type":"string",
                "example":""
              }
            }
          }
        }
      },
...
    },
...
    "securitySchemes": {
      "petstore_auth": {
        "type": "oauth2",
        "flows": {
          "implicit": {
            "authorizationUrl": "https://petstore3.swagger.io/oauth/authorize",
            "scopes": {
              "write:pets": "modify pets in your account",
              "read:pets": "read your pets"
            }
          }
        }
      },
      "api_key": {
        "type": "apiKey",
        "name": "api_key",
        "in": "header"
      }
    }
  }
}